pyproject.toml
[project]
name = "url-shrinker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.8",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
